fix(data): remove WebDataset io module and import collision
Changes:
- Move shared path validation and image extensions into index.py.
- Move tar-size accounting into pack.py and remove webdataset/io.py.
- Update loader and test imports; import BytesIO directly in tests.

Impact:
- Fix three test failures shared by the Ubuntu and Windows CI jobs.
- Reduce module count while preserving both helper implementations.

Verification:
- WebDataset tests and doctests: 149 passed, 6 skipped.
- WebDataset CLI tests: 5 passed.
- Full pre-commit and git diff --check: passed.
- Moved function ASTs unchanged; no stale webdataset.io references.

Residual limits:
- Local verification on macOS/Python 3.12; remote CI rerun pending.
- Test runs reported dependency warnings and six skips.

---
